Add settings to configure and disable encryption
This commit adds options to configure the encryption (cipher, hash, and
number of iterations) and to disable it all together. This way the
feature can be disabled in applications like the Erlang AMQP client or
Federation if it makes them unstable or crash (at the cost of losing the
credentials obfuscation). This seems like a reasonable trade-off while the
feature is being integrated in the system.

There is also now an encryption/decryption attempt on application startup to
crash early in case of a problem instead of crashing at an awkward moment,
e.g. when a Shovel or Federation link starts.

[#167149046]