Refactor code to supervisor / gen_server
Use of the library has changed so that end-users are required to set the secret at some point after the application is started. This allows RabbitMQ to set the secret to the cookie after distribution is started.

Removes the use of an ETS table in favor of gen_server state.