Implement fallback secret
Add possibility to set a secondary secret to be able to migrate between
multiple secret values.

API was kept to be compatible with previous behaviour, ie. strings are
converted to binaries.

Encrpyted values are now wrapped to be able to identify when a
decryption failed, as if the secret is invalid, it may succeed and
return junk data.